backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-env==1.1.3
gunicorn==21.2.0
psutil==5.9.6
openai==1.3.0
requests==2.31.0
google-generativeai==0.3.2
pytz==2023.3
PyJWT==2.8.0
stripe==7.8.0
razorpay==1.4.1
psycopg2-binary==2.9.9
redis==5.0.1
bcrypt==4.1.2
apscheduler
pytz